In process of changing types to use IntPtr
Former-commit-id: 6ac82e15f9e7c72d459ee28b93501319eb8077d9

+++ Anathema/GUI/Tools/Scanners/GUIChangeCounter.cs
@@ -61,7 +61,7 @@ namespace Anathema
         private void SetVariableSize()
         {
             Int32 VariableSize = (Int32)Math.Pow(2, VariableSizeTrackBar.Value);
-            VariableSizeValueLabel.Text = Conversions.BytesToMetric((UInt64)VariableSize);
+            VariableSizeValueLabel.Text = Conversions.BytesToMetric(VariableSize);
             ChangeCounterPresenter.SetVariableSize(VariableSize);
         }
